developer/lab/tools/RL Snake/requirements.txt
numpy===1.18.5
tensorflow===2.3.0
tensorflow_probability===0.11.1
pysdl2===0.9.7
gym===0.17.3
